// backend-v2/src/token/interfaces/sui-token.interface.ts
export interface SuiToken {
  type: string;
  objectId: string;
  name: string;
  supply: number;
  supplyInUsd: number;
  tokenPrice: number;
  dominance: number;
  circulatingSupply: number;
  marketCap: number;
  totalVolume: number;
  maxSupply: number;
  fdv: number;
  holders: number;
  denom: string;
  packageId: string;
  packageScamMessage: string | null;
  createTimestamp: number;
  creator: string;
  creatorName: string;
  creatorImg: string;
  creatorScamMessage: string | null;
  scamMessage: string | null;
  decimals: number;
  symbol: string;
  iconUrl: string;
  description: string;
  bridge: boolean;
  verified: boolean;
  afSupported: boolean;
  createdAt?: Date;
  updatedAt?: Date;
} 